fix(core): the features that have never worked on any box
Ten defects that all failed the same way: a type mismatch or a wrong function
name, converted by `.ok()` / `.unwrap_or(0)` into a plausible number, so nothing
ever surfaced. Each is small; the point is that together they are a list of
things the product claimed to do and never did.

  * SLEEP HAS ALWAYS READ "0.0 hours". `duration_minutes` is `integer`, decoded
    as `Option<i64>` — a decode error, `.ok()`d to None, for every sleep event
    ever recorded.
  * EVERY DATE-SCOPED SEARCH RETURNED NOTHING. The tool schema tells the model
    "ISO 8601, e.g. '2026-01-01'"; the code parsed with `parse_from_rfc3339`,
    which rejects a bare date. The Err became None became a bound NULL, and
    `occurred_at >= NULL` is never true — so the assistant reported "nothing in
    your record for that period" about a record that had it. Now parses both
    forms, and NULL means "no filter" rather than "match nothing", so a bad date
    widens the search instead of silently emptying it.
  * THE NARRATOR HAS NEVER SEEN A HEART RATE. `COUNT(*)` is bigint, read as i32,
    so `cnt` was always 0 and the `HEALTH:` line never entered a prompt.
  * RESTING HR WAS THE LITERAL 62.0 FOR EVERY OWNER. `AVG(integer)` returns
    `numeric` and `MIN(integer)` returns `integer`; both were decoded as f64 and
    both `.ok()?`d out to the hardcoded fallback — logged as if measured.
  * TRASH HAS NEVER BEEN VIEWABLE, and nothing deleted could be restored:
    `datetime('now','-30 days')` and two-argument `MAX(0, …)` are SQLite. On
    Postgres both are `function does not exist`. Verified against a live
    database before and after. Now `now() - interval '30 days'` and `GREATEST`.
  * THE STORAGE METER HAS NEVER COUNTED. `app_drive_usage`'s singleton row is
    created by no migration, no seed and no code path, so the UPDATE matched
    zero rows forever — silent, because an UPDATE that matches nothing is not an
    error. Now an upsert: a singleton whose absence disables a feature should be
    born the first time that feature is used, not seeded and hoped for.
  * /api/activity/metrics HAS NEVER RETURNED 200. `CAST(… AS INTEGER)` decoded
    as i64. Two sites; the two beside them already said BIGINT — half was fixed
    once.
  * COMPACTION HAS NEVER SUCCEEDED. `summary_up_to_index` and `summary_version`
    are bigint, read as i32 via `Row::get` — an unconditional panic, inside the
    SSE generator, so it killed the connection mid-turn with no error event and
    the chat stayed Critical forever.
  * DAY PAGES READ "0 new people, 0 new topics" on the day you met someone: two
    subqueries still named `r2.timestamp` and `e2.start_time` after the rename.
  * TOPIC NOVELTY returned max for everything when the baseline had days but no
    usable embeddings — the exact behavior the comment 140 lines above forbids
    ("'I have no baseline' and 'this is unprecedented' are opposite claims and
    must never produce the same number"). The caller already refuses to score in
    that case; the inner function had a second route to the old answer.

`api/census.rs` was already fixed by whoever renamed it; noting it so the list is
not read as complete when it was nine of ten.

The common thread, and the reason this is worth a commit message rather than a
line: eight of these are things `?` would have surfaced the first time anyone
opened the page. `.ok()` on a query result is not a style preference here — it is
what turned a rename into a decade of plausible zeros.

532 tests pass.

+++ virtues-core/src/dayline/context.rs
@@ -192,7 +192,7 @@ pub async fn build_hourly_context(
     .await
     {
         if let Some(r) = row {
-            let cnt: i32 = r.try_get("cnt").unwrap_or(0);
+            let cnt: i64 = r.try_get("cnt").unwrap_or(0);
             let avg: Option<f64> = r.try_get("avg_bpm").ok().flatten();
             if cnt > 0 {
                 if let Some(a) = avg {
